Network optimisations
*Replaced sfml with missing opt:: types
*Most packet data changed to Uint8

+++ PlaguedSurvival/NetworkOptimisations.hpp
@@ -3,10 +3,10 @@
 
 namespace opt
 {
-	typedef sf::Int32 ServerPacket;
-	typedef sf::Int32 ClientPacket;
-	typedef sf::Int32 PlayerIdentifier;
-	typedef sf::Int32 PlayerCount;
-	typedef sf::Int32 LocalPlayers;
-	typedef sf::Int32 Action;
+	typedef sf::Uint8 ServerPacket;
+	typedef sf::Uint8 ClientPacket;
+	typedef sf::Uint8 PlayerIdentifier;
+	typedef sf::Uint8 PlayerCount;
+	typedef sf::Uint8 LocalPlayers;
+	typedef sf::Uint8 Action;
 }
